+++ main/src/main/java/org/qbicc/main/Main.java
@@ -124,7 +124,7 @@ import org.qbicc.plugin.native_.NativeXtorLoweringHook;
 import org.qbicc.plugin.native_.PointerBasicBlockBuilder;
 import org.qbicc.plugin.native_.PointerTypeResolver;
 import org.qbicc.plugin.native_.StructMemberAccessBasicBlockBuilder;
-import org.qbicc.plugin.initializationcontrol.QbiccFeatureTypeBuilder;
+import org.qbicc.plugin.initializationcontrol.InitAtRuntimeTypeBuilder;
 import org.qbicc.plugin.objectmonitor.ObjectMonitorBasicBlockBuilder;
 import org.qbicc.plugin.opt.BlockParameterOptimizingVisitor;
 import org.qbicc.plugin.opt.FinalFieldLoadOptimizer;
@@ -151,7 +151,7 @@ import org.qbicc.plugin.reachability.ServiceLoaderAnalyzer;
 import org.qbicc.plugin.reflection.Reflection;
 import org.qbicc.plugin.reflection.ReflectionFactsSetup;
 import org.qbicc.plugin.reflection.ReflectionIntrinsics;
-import org.qbicc.plugin.reflection.ReflectiveMethodAccessorGenerator;
+import org.qbicc.plugin.reflection.ReflectiveElementTypeBuilder;
 import org.qbicc.plugin.reflection.VarHandleResolvingBasicBlockBuilder;
 import org.qbicc.plugin.serialization.BuildtimeHeap;
 import org.qbicc.plugin.serialization.ClassObjectSerializer;
@@ -390,8 +390,9 @@ public class Main implements Callable<DiagnosticContext> {
                                 builder.addTypeBuilderFactory(ThreadLocalTypeBuilder::new);
                                 builder.addTypeBuilderFactory(CoreAnnotationTypeBuilder::new);
                                 builder.addTypeBuilderFactory(ReachabilityAnnotationTypeBuilder::new);
+                                builder.addTypeBuilderFactory(ReflectiveElementTypeBuilder::new);
                                 builder.addTypeBuilderFactory(Patcher::getTypeBuilder);
-                                builder.addTypeBuilderFactory(QbiccFeatureTypeBuilder::new);
+                                builder.addTypeBuilderFactory(InitAtRuntimeTypeBuilder::new);
                                 builder.addTypeBuilderFactory(AccessorTypeBuilder::new);
 
                                 builder.setClassContextListener(Patcher::initialize);
@@ -412,6 +413,9 @@ public class Main implements Callable<DiagnosticContext> {
                                 builder.addPreHook(Phase.ADD, ReachabilityFactsSetup::setupAdd);
                                 builder.addPreHook(Phase.ADD, ReflectionFactsSetup::setupAdd);
                                 builder.addPreHook(Phase.ADD, ctxt -> SafePoints.selectStrategy(ctxt, nogc ? SafePoints.Strategy.NONE : SafePoints.Strategy.GLOBAL_FLAG));
+                                builder.addPreHook(Phase.ADD, compilationContext -> {
+                                    QbiccFeatureProcessor.process(compilationContext, qbiccYamlFeatures, qbiccFeatures);
+                                });
                                 if (llvm) {
                                     builder.addPreHook(Phase.ADD, LLVMIntrinsics::register);
                                 }
@@ -422,9 +426,6 @@ public class Main implements Callable<DiagnosticContext> {
                                 builder.addPreHook(Phase.ADD, Reflection::get);
                                 builder.addPreHook(Phase.ADD, UnwindExceptionStrategy::get);
                                 builder.addPreHook(Phase.ADD, GcCommon::registerIntrinsics);
-                                builder.addPreHook(Phase.ADD, compilationContext -> {
-                                    QbiccFeatureProcessor.process(compilationContext, qbiccYamlFeatures, qbiccFeatures);
-                                });
                                 builder.addPreHook(Phase.ADD, compilationContext -> {
                                     Vm vm = compilationContext.getVm();
                                     VmThread initThread = vm.newThread("initialization", vm.getMainThreadGroup(), false,  Thread.currentThread().getPriority());
@@ -465,7 +466,6 @@ public class Main implements Callable<DiagnosticContext> {
                                     .andThen(new BuildTimeOnlyElementHandler())
                                     .andThen(new ElementInitializer())));
                                 builder.addPreHook(Phase.ADD, new ElementReachableAdapter(ReachabilityInfo::processReachableElement));
-                                builder.addPreHook(Phase.ADD, new ElementReachableAdapter(new ReflectiveMethodAccessorGenerator()));
                                 builder.addPreHook(Phase.ADD, new ElementReachableAdapter(new ElementVisitorAdapter(new DotGenerator(Phase.ADD, graphGenConfig))));
                                 builder.addBuilderFactory(Phase.ADD, BuilderStage.TRANSFORM, IntrinsicBasicBlockBuilder::createForAddPhase);
                                 if (nogc) {
@@ -496,12 +496,6 @@ public class Main implements Callable<DiagnosticContext> {
                                 builder.addBuilderFactory(Phase.ADD, BuilderStage.INTEGRITY, DeferenceBasicBlockBuilder::new);
                                 builder.addBuilderFactory(Phase.ADD, BuilderStage.INTEGRITY, ReachabilityBlockBuilder::new);
                                 builder.addBuilderFactory(Phase.ADD, BuilderStage.INTEGRITY, StaticChecksBasicBlockBuilder::new);
-                                builder.addPostHook(Phase.ADD, ctxt -> {
-                                    Vm vm = ctxt.getVm();
-                                    vm.doAttached(vm.newThread("FieldAccessor Generation", vm.getMainThreadGroup(), false, Thread.currentThread().getPriority()), () -> {
-                                        Reflection.get(ctxt).generateFieldAccessors();
-                                    });
-                                });
                                 builder.addPostHook(Phase.ADD, ctxt -> {
                                     Vm vm = ctxt.getVm();
                                     vm.doAttached(vm.newThread("ReflectionData Generation", vm.getMainThreadGroup(), false, Thread.currentThread().getPriority()), () -> {
