+++ plugins/initialization-control/pom.xml
@@ -30,11 +30,14 @@
             <artifactId>qbicc-plugin-reachability</artifactId>
         </dependency>
         <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>qbicc-plugin-reflection</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
             <version>2.14.2</version>
         </dependency>
-
     </dependencies>
 
 </project>
