+++ plugins/reachability/src/main/java/org/qbicc/plugin/reachability/ReachabilityRoots.java
@@ -1,16 +1,12 @@
 package org.qbicc.plugin.reachability;
 
-import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.qbicc.context.AttachmentKey;
 import org.qbicc.context.CompilationContext;
-import org.qbicc.type.definition.LoadedTypeDefinition;
-import org.qbicc.type.definition.element.ConstructorElement;
 import org.qbicc.type.definition.element.ExecutableElement;
 import org.qbicc.type.definition.element.FieldElement;
-import org.qbicc.type.definition.element.MethodElement;
 import org.qbicc.type.definition.element.StaticFieldElement;
 
 /**
@@ -31,9 +27,8 @@ public class ReachabilityRoots {
     private static final AttachmentKey<ReachabilityRoots> KEY = new AttachmentKey<>();
 
     private final CompilationContext ctxt;
-    private final Set<LoadedTypeDefinition> reflectiveClasses = ConcurrentHashMap.newKeySet();
     private final Set<ExecutableElement> reflectiveMethods = ConcurrentHashMap.newKeySet();
-    private final Set<FieldElement> reflectiveFields = ConcurrentHashMap.newKeySet();
+    private final Set<StaticFieldElement> heapRoots = ConcurrentHashMap.newKeySet();
     private final Set<ExecutableElement> autoQueuedMethods = ConcurrentHashMap.newKeySet();
     private final Set<ExecutableElement> dispatchTableMethods = ConcurrentHashMap.newKeySet();
 
@@ -55,24 +50,11 @@ public class ReachabilityRoots {
 
     public void reportStats() {
         ReachabilityInfo.LOGGER.debugf("  Auto-queued methods:           %s", autoQueuedMethods.size());
-        ReachabilityInfo.LOGGER.debugf("  Reflectively accessed classes: %s", reflectiveClasses.size());
         ReachabilityInfo.LOGGER.debugf("  Reflectively accessed methods: %s", reflectiveMethods.size());
-        ReachabilityInfo.LOGGER.debugf("  Reflectively accessed fields:  %s", reflectiveFields.size());
+        ReachabilityInfo.LOGGER.debugf("  Reflective heap roots (statics): %s", heapRoots.size());
     }
 
-    public void registerReflectiveClass(LoadedTypeDefinition ltd) {
-        reflectiveClasses.add(ltd);
-    }
-
-    public boolean registerReflectiveMethod(MethodElement e) {
-        boolean added = reflectiveMethods.add(e);
-        if (added) {
-            ctxt.enqueue(e);
-        }
-        return added;
-    }
-
-    public boolean registerReflectiveConstructor(ConstructorElement e) {
+    public boolean registerReflectiveEntrypoint(ExecutableElement e) {
         boolean added = reflectiveMethods.add(e);
         if (added) {
             ctxt.enqueue(e);
@@ -92,8 +74,8 @@ public class ReachabilityRoots {
         return dispatchTableMethods.add(e);
     }
 
-    public boolean registerReflectiveField(FieldElement f) {
-        return reflectiveFields.add(f);
+    public boolean registerHeapRoot(StaticFieldElement f) {
+        return heapRoots.add(f);
     }
 
     // During Analyze, we can allow reachability analysis to figure things out for us.
@@ -103,16 +85,11 @@ public class ReachabilityRoots {
         for (ExecutableElement e : roots.autoQueuedMethods) {
             info.processRootReachableElement(e);
         }
-        for (LoadedTypeDefinition ltd : roots.reflectiveClasses) {
-            info.getAnalysis().processReachableType(ltd, null);
-        }
         for (ExecutableElement e : roots.reflectiveMethods) {
             info.processRootReachableElement(e);
         }
-        for (FieldElement f : roots.reflectiveFields) {
-            if (f.isStatic()) {
-                info.getAnalysis().processReachableStaticFieldAccess((StaticFieldElement) f, null);
-            }
+        for (StaticFieldElement f : roots.heapRoots) {
+            info.getAnalysis().processReachableStaticFieldAccess(f, null);
         }
     }
 
@@ -134,21 +111,5 @@ public class ReachabilityRoots {
         for (ExecutableElement e : roots.dispatchTableMethods) {
             ctxt.enqueue(e);
         }
-    }
-
-    public boolean isReflectiveClass(LoadedTypeDefinition ltd) {
-        return reflectiveClasses.contains(ltd);
-    }
-
-    public Set<LoadedTypeDefinition> getReflectiveClasses() {
-        return new HashSet<>(reflectiveClasses);
-    }
-
-    public Set<FieldElement> getReflectiveFields() {
-        return new HashSet<>(reflectiveFields);
-    }
-
-    public Set<ExecutableElement> getReflectiveMethods() {
-        return new HashSet<>(reflectiveMethods);
     }
 }
